Revert "add change request"
This reverts commit 87cb3d435d5f96563c6d9e1290c3c0a239f7aaf0.

+++ CB-Backend/src/main/java/edu/ucmo/cbbackend/model/ChangeRequest.java
@@ -1,10 +1,7 @@
 package edu.ucmo.cbbackend.model;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.io.Serializable;
 
@@ -32,7 +29,7 @@ public class ChangeRequest implements Serializable {
     private ChangeType changeType;
 
     @Column(nullable = false)
-    private Long applicationId;
+    private Integer applicationId;
 
     @Column(nullable = false)
     private String description;
